added rad2deg and deg2rad func to utils. fixed render.py and render sub in sdfsMod.f90. Added exp setup to subs.f90. added dynamic selection of sdf layer for each packet. fixed bug in intaus packet layer handling.

+++ render.py
@@ -30,7 +30,7 @@ parser.add_argument("-f", "--file", type=str,
                     help="Name of file to be plotted.")
 parser.add_argument("-n", "--ndim", type=int, default=200,
                     help="Gives the dimensions of the data cube to be examined.")
-parser.add_argument("-l", "--levels", type=int, default=0,
+parser.add_argument("-l", "--levels", type=np.float32, default=0.0,
                     help="Isolevel to render a contour on.")
 
 args = parser.parse_args()
